requirements.txt
argparse
requests
python-dotenv
pandas
tqdm